% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{americas_gdp_pc}
\alias{americas_gdp_pc}
\title{GDP per capita PPP in America}
\description{
GDP per capita with purchasing power parity data, for the countries of America, 1996-1998. Data from the World Bank's World Development Indicators.
}
\references{
\url{https://data.worldbank.org/indicator/NY.GDP.PCAP.PP.CD}
}
\keyword{data}
