% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{approval_wide2}
\alias{approval_wide2}
\title{Presidential approval  (wide 2)}
\description{
Presidential approval data by quarter, 2000-2014, compiled by Catherine Reyes-Housholder. Some variables from the Quality of Government project (2017) are added. Wide format (2).
}
\references{
Reyes-Housholder, C. (2019). A Theory of Gender’s Role on Presidential Approval Ratings in Corrupt Times. Political Research Quarterly.

\url{https://qog.pol.gu.se/data}
}
\keyword{data}
