% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{cosponsorship_arg}
\alias{cosponsorship_arg}
\title{Cosponsorship in the Argentine Senate (1983)}
\description{
Data from the joint presentation of bills in the Argentine Senate
}
\references{
Alemán, E., Calvo, E., Jones, M. P., & Kaplan, N. (2009). Comparing Cosponsorship and Roll‐Call Ideal Points. Legislative Studies Quarterly, 34(1), 87-116.
}
\keyword{data}
