% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{dem_breakdown}
\alias{dem_breakdown}
\title{Democratic breakdowns in Latin America}
\description{
Democratic breakdowns in Latin America data, 1900-2010, by Mainwaring & Pérez-Liñán (2013).
}
\references{
Mainwaring, S., & Pérez-Liñán, A. (2013). Democracies and dictatorships in Latin America: emergence, survival, and fall. Cambridge University Press.
}
\keyword{data}
