% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{direct_democracy}
\alias{direct_democracy}
\title{Direct democracy}
\description{
Data from institutions of direct democracy in the world, compiled by David Altman (2018).
}
\references{
Altman, D. (2018). Citizenship and Contemporary Direct Democracy. Cambridge University Press.
}
\keyword{data}
