% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{edad_educ}
\alias{edad_educ}
\title{Data from the states of Brazil}
\description{
Data from the states of Brazil, compiled by Freire (2018)
}
\references{
Freire, D. (2018). Evaluating the effect of homicide prevention strategies in são paulo, brazil: A synthetic control approach. Latin American Research Review, 53(2), 231. \url{https://doi.org/10.25222/larr.334}
}
\keyword{data}
