% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{edu_age}
\alias{edu_age}
\title{Age and education as confounders}
\description{
Simulated data, in which age and education are confounders for the treatment of using mosquito nets. Created by Andrew Heiss.
}
\keyword{data}
