% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{lapop}
\alias{lapop}
\title{LAPOP survey}
\description{
LAPOP survey's data (2016)
}
\references{
\url{https://www.vanderbilt.edu/lapop/ab2016/AB2017-v18.0-Spa-170202_W.pdf}
}
\keyword{data}
