% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{latinobarometer}
\alias{latinobarometer}
\title{Latinobarometer}
\description{
Data subset of the Latinobarometer survey, 2018.
}
\references{
Latinobarómetro (2018). Latinobarómetro: Banco de datos. \url{https://www.latinobarometro.org/latContents.jsp}
}
\keyword{data}
