% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{messy_summary}
\alias{messy_summary}
\title{Messy summary of American countries GDP per capita}
\description{
"Messy" summary of GDP per capita of American countries, in which the names of the countries don't follow a standard. Per capita GDP data from the World Bank's World Development Indicators.
}
\references{
\url{https://data.worldbank.org/indicator/NY.GDP.PCAP.PP.CD}
}
\keyword{data}
