% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{mosquito_nets}
\alias{mosquito_nets}
\title{Mosquito nets}
\description{
Simulated data, in which the treatment of mosquito nets is evaluated on the risk of malaria. Created by Andrew Heiss.
}
\keyword{data}
