% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{municipal_data}
\alias{municipal_data}
\title{Chilean municipalities data (1992-2012)}
\description{
Dataframe builded by Soledad Araya, with variables for Chilean municipalities between 1992 and 2012.
}
\keyword{data}
