% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{poltweets}
\alias{poltweets}
\title{Political tweets in Chile}
\description{
Tweets from Chilean deputies, between April and June 2018. Compiled by Sebastián Huneeus.
}
\references{
\url{https://www.vanderbilt.edu/lapop/ab2016/AB2017-v18.0-Spa-170202_W.pdf}
}
\keyword{data}
