% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{senators_arg}
\alias{senators_arg}
\title{Senators of Argentina (1983)}
\description{
Argentine senators data, 1983.
}
\references{
Alemán, E., Calvo, E., Jones, M. P., & Kaplan, N. (2009). Comparing Cosponsorship and Roll‐Call Ideal Points. Legislative Studies Quarterly, 34(1), 87-116.
}
\keyword{data}
