% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{treaties}
\alias{treaties}
\title{Treaties}
\description{
Data of adoption, ratification, etc. of international treaties. Compiled by Carsten Schulz based on the United Nations repository.
}
\references{
\url{https://www.carstenschulz.eu/}

\url{https://tratados.un.org/}
}
\keyword{data}
