% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{us_latam}
\alias{us_latam}
\title{Political proximity between the US and Latin American countries}
\description{
Data of political proximity between the US and Latin American countries, generated from their votes in the United Nations General Assembly. Original data from Voeten (2012).
}
\references{
Voeten, E. (2012). Data and analyses of voting in the un general assembly (SSRN Scholarly Paper ID 2111149). Social Science Research Network. \url{https://papers.ssrn.com/abstract=2111149}
}
\keyword{data}
