% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{welfare}
\alias{welfare}
\title{Welfare in Latin America}
\description{
Welfare data in Latin America, by Huber & Stephens (2012)
}
\references{
\url{http://huberandstephens.web.unc.edu/common-works/data/}
}
\keyword{data}
